requirements.txt
Flask
gunicorn
pymongo
python-dotenv
beautifulsoup4
requests
playwright
numpy
scikit-learn
torch
transformers